Move tcli tests to tcli package
