+++ ci/integration/integration_test.go
@@ -43,78 +43,6 @@ func init() {
 	}
 }
 
-func TestTCli(t *testing.T) {
-	t.Parallel()
-	ctx := context.Background()
-
-	container, err := tcli.NewContainerRunner(ctx, &tcli.ContainerConfig{
-		Image: "ubuntu:latest",
-		Name:  "test-container",
-		BindMounts: map[string]string{
-			binpath: "/bin/coder",
-		},
-	})
-	assert.Success(t, "new run container", err)
-	defer container.Close()
-
-	container.Run(ctx, "echo testing").Assert(t,
-		tcli.Success(),
-		tcli.StderrEmpty(),
-		tcli.StdoutMatches("esting"),
-	)
-
-	container.Run(ctx, "sleep 1.5 && echo 1>&2 stderr-message").Assert(t,
-		tcli.Success(),
-		tcli.StdoutEmpty(),
-		tcli.StderrMatches("message"),
-		tcli.DurationGreaterThan(time.Second),
-	)
-
-	cmd := exec.CommandContext(ctx, "cat")
-	cmd.Stdin = strings.NewReader("testing")
-
-	container.RunCmd(cmd).Assert(t,
-		tcli.Success(),
-		tcli.StderrEmpty(),
-		tcli.StdoutMatches("testing"),
-	)
-
-	container.Run(ctx, "which coder").Assert(t,
-		tcli.Success(),
-		tcli.StdoutMatches("/bin/coder"),
-		tcli.StderrEmpty(),
-	)
-
-	container.Run(ctx, "coder version").Assert(t,
-		tcli.StderrEmpty(),
-		tcli.Success(),
-		tcli.StdoutMatches("linux"),
-	)
-}
-
-func TestHostRunner(t *testing.T) {
-	t.Parallel()
-	var (
-		c   tcli.HostRunner
-		ctx = context.Background()
-	)
-
-	c.Run(ctx, "echo testing").Assert(t,
-		tcli.Success(),
-		tcli.StderrEmpty(),
-		tcli.StdoutMatches("testing"),
-	)
-
-	wd, err := os.Getwd()
-	assert.Success(t, "get working dir", err)
-
-	c.Run(ctx, "pwd").Assert(t,
-		tcli.Success(),
-		tcli.StderrEmpty(),
-		tcli.StdoutMatches(wd),
-	)
-}
-
 func TestCoderCLI(t *testing.T) {
 	t.Parallel()
 	ctx, cancel := context.WithTimeout(context.Background(), time.Minute*5)
@@ -129,6 +57,12 @@ func TestCoderCLI(t *testing.T) {
 	})
 	assert.Success(t, "new run container", err)
 	defer c.Close()
+
+	c.Run(ctx, "which coder").Assert(t,
+		tcli.Success(),
+		tcli.StdoutMatches("/usr/sbin/coder"),
+		tcli.StderrEmpty(),
+	)
 
 	c.Run(ctx, "coder version").Assert(t,
 		tcli.StderrEmpty(),
